feat(reddit/trending): cache fix B — move searchParams to client island (matches #1606)
The /reddit/trending RSC awaited `searchParams` for an SSR empty-tab redirect,
forcing full-dynamic rendering and rendering `revalidate=1800` inert. Mirrors
the /agent-commerce regression resolved in #1606.

Changes:
- Remove `searchParams` Promise prop + `await searchParams` from page.tsx.
- Delete the SSR empty-tab redirect block (43 lines). The same redirect
  already runs client-side in AllTrendingTabs via useEffect+router.replace
  (lines 306-310). Trades the SSR pre-hydration flash for ISR cacheability.
- New thin client island `RedditTrendingFilteredContent` wraps the existing
  `<Suspense><AllTrendingTabs/></Suspense>` block. Matches the agent-commerce
  client-island pattern so the boundary is explicit and the page shell stays
  statically prerenderable.

Verification:
- `npm run typecheck` — clean (no diagnostics).
- `npm run lint:guards` — 0 missing budgets / 0 orphaned configs.

Preserved:
- All filter logic in AllTrendingTabs (unchanged).
- `export const revalidate = 1800` on the RSC.
- Bundled-JSON SSR fallback + cold-state EmptyState rendering.
- KPI band, InventoryBand, FreshnessBadge wiring.
